requirements.txt
awscli==1.40.35
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.36
botocore==1.38.36
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
docutils==0.19
filelock==3.18.0
Flask==3.1.1
Flask-PyMongo==3.0.1
gdown==5.2.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pyasn1==0.6.1
pymongo==4.13.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.7.2
s3transfer==0.13.0
six==1.17.0
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3